pyproject.toml
[project]
name = "ultrabot-ai"
version = "0.1.0"
description = "A robust, feature-rich personal AI assistant framework with circuit breakers, failover, parallel tools, and plugin system"
readme = { file = "README.md", content-type = "text/markdown" }
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "ultrabot contributors"}
]
keywords = ["ai", "agent", "chatbot", "assistant", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "typer>=0.20.0,<1.0.0",
    "anthropic>=0.45.0,<1.0.0",
    "openai>=2.8.0",
    "pydantic>=2.12.0,<3.0.0",
    "pydantic-settings>=2.12.0,<3.0.0",
    "httpx>=0.28.0,<1.0.0",
    "loguru>=0.7.3,<1.0.0",
    "rich>=14.0.0,<15.0.0",
    "prompt-toolkit>=3.0.50,<4.0.0",
    "questionary>=2.0.0,<3.0.0",
    "croniter>=6.0.0,<7.0.0",
    "tiktoken>=0.12.0,<1.0.0",
    "aiosqlite>=0.21.0,<1.0.0",
    "json-repair>=0.57.0,<1.0.0",
    "chardet>=3.0.2,<6.0.0",
    "ddgs>=9.5.5,<10.0.0",
    "websockets>=16.0,<17.0",
]

[project.optional-dependencies]
telegram = [
    "python-telegram-bot[socks]>=22.6,<23.0",
]
discord = [
    "discord.py>=2.4.0,<3.0.0",
]
slack = [
    "slack-sdk>=3.39.0,<4.0.0",
    "slackify-markdown>=0.2.0,<1.0.0",
]
mcp = [
    "mcp>=1.26.0,<2.0.0",
]
webui = [
    "fastapi>=0.115.0,<1.0.0",
    "uvicorn[standard]>=0.34.0,<1.0.0",
]
all = [
    "ultrabot-ai[telegram,discord,slack,mcp,webui]",
]
dev = [
    "pytest>=9.0.0,<10.0.0",
    "pytest-asyncio>=1.3.0,<2.0.0",
    "pytest-cov>=6.0.0,<7.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
ultrabot = "ultrabot.cli.commands:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
include = [
    "ultrabot/**/*.py",
    "ultrabot/templates/**/*.md",
    "ultrabot/skills/**/*.md",
    "ultrabot/webui/static/**/*",
]

[tool.hatch.build.targets.wheel]
packages = ["ultrabot"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.coverage.run]
source = ["ultrabot"]
omit = ["tests/*", "**/tests/*"]
